requirements.txt
openai
agno
pydantic
sqlalchemy
googlesearch-python
pycountry
python-dotenv
lxml_html_clean
lancedb
tantivy
matplotlib
networkx
pygraphviz